Refactor scripts
In separating the install and scriptopoly scripts there was a need to
change the commands for both the 'Poetry scripts' and 'setup.py
console_scripts'. This also resulted in needing two script directories,
one inside the other. More info on why that is can be found in a
comment in 'scripts.scripts.__init__'.

This also prevents installing Monopoly-Probabilities if the current
working directory is not 'Monopoly-Probabilities'. This helps avoid
various errors including setup.py failing because it can't be found and
our 'scripts.pth' file being constructed incorrectly.
