--- script/__init__.py
@@ -1,1 +1,0 @@
-from script.script import main
